\name{shashirojanlab3-package}
\alias{shashirojanlab3-package}
\alias{shashirojanlab3}
\docType{package}
\title{
\packageTitle{shashirojanlab3}
}
\description{
\packageDescription{shashirojanlab3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{shashirojanlab3}
\packageIndices{shashirojanlab3}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{shashirojanlab3}

Maintainer: \packageMaintainer{shashirojanlab3}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
