\name{ShashiRojanlab3-package}
\alias{ShashiRojanlab3-package}
\alias{ShashiRojanlab3}
\docType{package}
\title{
\packageTitle{ShashiRojanlab3}
}
\description{
\packageDescription{ShashiRojanlab3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ShashiRojanlab3}
\packageIndices{ShashiRojanlab3}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ShashiRojanlab3}

Maintainer: \packageMaintainer{ShashiRojanlab3}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
